feat(vs2022_third): C4244 fix conversions in NanoSpecDoc and AlarmLogDlg — 2 files
Fix 12 C4244 warnings by adding explicit static_cast<long>:
- NanoSpecDoc.cpp lines 1050-1055: Cast double→long for coordinate calculations (6 warnings)
- AlarmLogDlg.cpp lines 1050-1055: Cast double→long for coordinate calculations (6 warnings)

No behavior changes, only explicit type conversions.

+++ SRC/NanoSpec/NanoSpecDoc.cpp
@@ -1052,7 +1052,7 @@ void CNanoSpecDoc::SetStartDateTime(SYSTEMTIME* syStartDateTime)
 //	‘ª’èI—¹“úŽž‚ÌŽæ“¾
 void CNanoSpecDoc::GetEndDateTime(SYSTEMTIME* syEndDateTime)
 {
-	*syEndDateTime = m_pMeasureddata->syEndDateTime;
+	*syEndDateTime = static_cast<long>(m_pMeasureddata->syEndDateTime);
 }
 
 /////////////////////////////////////////////////////////////////////////////
